IB/core: Change lid size in struct ib_port_attr from 16 to 32 bits
OPA devices can support larger number of nodes than 16 bit lids
can accommodate. This change increases the lid size from 16 to
32 bits in the kernel so kernel clients such as ipoib and srp
can get and set 32 bit lids in port attribute.
UAPI is unchanged and is still only 16 bits.

+++ drivers/infiniband/ulp/ipoib/ipoib_multicast.c
@@ -590,7 +590,10 @@ void ipoib_mcast_join_task(struct work_struct *work)
 			  port_attr.state);
 		return;
 	}
-	priv->local_lid = port_attr.lid;
+	/* TODO: This casting will not be needed once priv->local_lid
+	 * is extended to 32 bits.
+	 */
+	priv->local_lid = (u16)port_attr.lid;
 
 	if (ib_query_gid(priv->ca, priv->port, 0, &priv->local_gid, NULL))
 		ipoib_warn(priv, "ib_query_gid() failed\n");
